requirements.txt
cycler==0.10.0
grad-cam==1.2.9
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
threadpoolctl==2.2.0
timm==0.4.12
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.2
ttach==0.0.3
typing-extensions==3.10.0.0
